ci: replace stripped-binary audit with QEMU wheel smoke
The AVX-512 portion of tools/ci/audit_x86_build.py never ran on the
actual published wheel (release.yml didn't call it), and on stripped
binaries -- which is what cibuildwheel produces -- it reports false
positives: objdump's symbol labels stretch past symbol-size boundaries
into adjacent unlabeled code, so AVX-512 instructions in template
instantiations get attributed to whichever exported symbol happened to
precede them. Verified on the v0.4.1 wheel: the audit "fails", but
qemu-user smoke under Haswell and Nehalem shows the wheel actually
dispatches correctly and never SIGILLs.

The compile-flag half of the audit is still useful and low-noise:

* Renames tools/ci/audit_x86_build.py to tools/ci/audit_build_flags.py
  and trims it to the audit_compile_commands check. This is what
  caught the libstim -march=native bug (#95).

Adds the runtime check the audit was supposed to provide:

* New .github/scripts/wheel_smoke.sh runs a tiny clifft program under
  qemu-x86_64 with a chosen -cpu model. Takes (cpu_model, force_isa,
  expected) so each matrix row is an explicit assertion. The "fail"
  expectation requires a clean RuntimeError from the CLIFFT_FORCE_ISA
  trap path and rejects SIGILL / exit 132 explicitly.
* ci.yml release-cpp-smoke now installs the release-baseline Python
  package, qemu-user, and runs five smoke rows on every PR:
  Haswell/auto, Haswell/avx2, Haswell/avx512 (clean trap),
  Nehalem/auto, Nehalem/avx2 (clean trap). Cost ~70s.
* release.yml installs the just-built Linux x86_64 wheel into a fresh
  venv and runs the same matrix against the actual cibuildwheel
  artifact. Skipped on other arches/OSes (qemu-user-mode is
  x86_64-specific, and the other matrix entries use the generic
  baseline anyway).

Verified locally against the v0.4.1 PyPI wheel: all five matrix rows
behave as expected on Haswell and Nehalem emulation.
